+++ src/services/ProdutoService.py
@@ -5,31 +5,41 @@ from repositories import ProdutoRepository as repository;
 from models.Produto import ProdutoModel;
 from entities.ProdutoEntity import ProdutoDB; 
 
+from services.exceptions.DatabaseException import DatabaseException;
+from services.exceptions.ResourceNotFoundException import ResourceNotFoundException;
+
 def insert(newProduto: ProdutoModel):
     try:
         produtoDB = ProdutoDB(None, newProduto.nome, newProduto.descricao, newProduto.foto, newProduto.valor_unitario);
         repository.insert(produtoDB)
         return newProduto;
-    except Exception as e:
-        raise e;
+    except DatabaseException as databaseException:
+        raise databaseException;
         
 def findAll():
     try:
         produtos = repository.findAll();
+        if len(produtos) == 0:
+            raise ResourceNotFoundException("Nenhum objeto encontrado.");
         return produtos;
-    except Exception as e:
-        raise e;
+    except DatabaseException as databaseException:
+        raise databaseException;
         
 def findById(id: int):
     try:
         produto = repository.findById(id);
+        if produto is None:
+            raise ResourceNotFoundException(f"Objeto não encontrado com o ID. {id}");
         return produto;
-    except Exception as e:
-        raise e;
+    except DatabaseException as databaseException:
+        raise databaseException;
         
 def update(id: int, newProduto: ProdutoModel):
     try:
         produto = repository.findById(id);
+        
+        if produto is None:
+            raise ResourceNotFoundException(f"Objeto não encontrado com o ID. {id}");
         
         produto.nome = newProduto.nome;
         produto.descricao = newProduto.descricao;
@@ -38,12 +48,16 @@ def update(id: int, newProduto: ProdutoModel):
         
         repository.update(produto)
         return produto;
-    except Exception as e:
-        raise e;
+    except DatabaseException as databaseException:
+        raise databaseException;
 
 def delete(id: int):
     try:
         produto = findById(id);
+        
+        if produto is None:
+            raise ResourceNotFoundException(f"Objeto não encontrado com o ID. {id}");
+        
         repository.delete(produto);
-    except Exception as e:
-        raise e;
+    except DatabaseException as databaseException:
+        raise databaseException;
